# DROOLS-6725 DMN DMNLiteralExpressionEvaluator to internalize FEEL